requirements.txt
psycopg2
Flask-Script
flask_migrate
rq
rq_scheduler
Flask-RQ2
flask-redis
Flask-Mail
requests
ipdb
gunicorn
